* ssize_t should be 32bits on WIN32 [rtj]

+++ src/tirpc/rpc/types.h
@@ -49,9 +49,9 @@ typedef uint8_t u_int8_t;
 typedef uint16_t u_int16_t;
 typedef uint32_t u_int32_t;
 typedef uint64_t u_int64_t;
-typedef unsigned long long int u_quad_t;
-typedef long long int quad_t;
-typedef long long int ssize_t;
+typedef unsigned long long u_quad_t;
+typedef long long quad_t;
+typedef long ssize_t;
 typedef uint64_t caddr_t;
 #endif
 #ifndef _FDOPEN
